+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - add `View calendar` to `scheduled_jobs` where all scheduled jobs can be shown in a calendar for a user
  - add button `View Action` on scheduled job view, if the job was scheduled from an action. The button leads to the action that scheduled the job
  - in style `register` - now for group we can select multiple groups which will be assigned to the user
+ - separate `groups` in `groups` and `db_roles`. The groups can give access to `experimental` pages and can be used in conditions. DB roles are used to give privileges and are used for management tool.
 
 # v6.0.0
 ###  New features